requirements.txt
pytest==7.2.0
allure-pytest==2.12.0
requests==2.28.1
voluptuous==0.13.1
pytest-voluptuous==1.2.0
curlify
